separate functions/ to NEW event_function.ex from FAT event_controller.ex
